improve: remove moka cache thread pool * improve: cli add use statement
