+++ client/src/command.rs
@@ -1,6 +1,7 @@
 //! Command within CLI
 
 use crate::ctx::{ResultSet, SessionContext};
+use crate::exec::connect_database;
 use crate::functions::{display_all_functions, Function};
 use crate::print_format::PrintFormat;
 use crate::print_options::PrintOptions;
@@ -33,11 +34,6 @@ pub enum OutputFormat {
     ChangeFormat(String),
 }
 
-fn is_system_table_db(db: &str) -> bool {
-    let db = db.to_ascii_lowercase();
-    db.eq("cluster_schema") || db.eq("information_schema")
-}
-
 impl Command {
     pub async fn execute(
         &self,
@@ -47,22 +43,10 @@ impl Command {
         let now = Instant::now();
         match self {
             Self::Help => print_options.print_batches(&all_commands_info(), now),
-            Self::ConnectDatabase(database) => {
-                if is_system_table_db(database) {
-                    ctx.set_database(database);
-                    return Ok(());
-                }
-                let old_database = ctx.get_database().to_string();
-                ctx.set_database(database);
-                match ctx.sql(format!("DESCRIBE DATABASE {}", database)).await {
-                    Ok(_) => Ok(()),
-                    Err(e) => {
-                        println!("Cannot connect to database {}.", database);
-                        ctx.set_database(old_database.as_str());
-                        Err(e)
-                    }
-                }
-            }
+            Self::ConnectDatabase(database) => connect_database(database, ctx).await.map_err(|e| {
+                println!("Cannot connect to database {}.", database);
+                e
+            }),
             Self::ListTables => {
                 let results = ctx.sql("SHOW TABLES".to_string()).await?;
                 print_options.print_batches(&results, now)
